Move ModelStub in AddSessionTest to class ModelStub to repeat less code
